telecomlib/smpp/requirements.txt
python3-smpplib==1.0.22
